+++ src/Mvc/Controller/KyteSectionTemplateController.php
@@ -15,7 +15,6 @@ class KyteSectionTemplateController extends ModelController
                 $r['html'] = bzcompress($r['html'], 9);
                 $r['stylesheet'] = bzcompress($r['stylesheet'], 9);
                 $r['javascript'] = bzcompress($r['javascript'], 9);
-                $r['javascript_obfuscated'] = bzcompress($r['javascript_obfuscated'], 9);
                 $r['block_layout'] = bzcompress($r['block_layout'], 9);
                 break;
 
@@ -32,7 +31,6 @@ class KyteSectionTemplateController extends ModelController
                 $r['html'] = bzdecompress($r['html']);
                 $r['stylesheet'] = bzdecompress($r['stylesheet']);
                 $r['javascript'] = bzdecompress($r['javascript']);
-                $r['javascript_obfuscated'] = bzdecompress($r['javascript_obfuscated']);
                 $r['block_layout'] = bzdecompress($r['block_layout']);
                 break;