+++ src/Mvc/Model/KytePage.php
@@ -142,15 +142,6 @@ $KytePage = [
 			'date'		=> false,
 		],
 
-		'obfuscate_js' => [
-			'type'		=> 'i',
-			'required'	=> false,
-			'size'		=> 1,
-			'unsigned'	=> true,
-			'default'	=> 0,
-			'date'		=> false,
-		],
-
 		'is_js_module' => [
 			'type'		=> 'i',
 			'required'	=> false,